+++ examples/terminal_logging/src/main.rs
@@ -2,7 +2,9 @@ use handsome_logger::{ColorChoice, Config, TermLogger, TerminalMode};
 use log::*;
 
 fn main() {
-    TermLogger::init(Config::default(), TerminalMode::Mixed, ColorChoice::Auto).unwrap();
+    handsome_logger::init();
+    // or longer version which do same thing:
+    // TermLogger::init(Config::default(), TerminalMode::Mixed, ColorChoice::Auto).unwrap();
 
     trace!("Got TRACE");
     debug!("Got DEBUG");